cmd/compile: more lineno type conversions
Reviewed-on: https://go-review.googlesource.com/20156
Run-TryBot: Matthew Dempsky <mdempsky@google.com>
TryBot-Result: Gobot Gobot <gobot@golang.org>

+++ src/cmd/compile/internal/gc/align.go
@@ -125,7 +125,7 @@ func dowidth(t *Type) {
 	if t.Width == -2 {
 		if !t.Broke {
 			t.Broke = true
-			yyerrorl(int32(t.Lineno), "invalid recursive type %v", t)
+			yyerrorl(t.Lineno, "invalid recursive type %v", t)
 		}
 
 		t.Width = 0
@@ -142,7 +142,7 @@ func dowidth(t *Type) {
 	defercalc++
 
 	lno := lineno
-	lineno = int32(t.Lineno)
+	lineno = t.Lineno
 	t.Width = -2
 	t.Align = 0
